Rename dateInput2MigrationUtils -> dateInputMigrationUtils
